New Generate Button | Structure Refactor
